+++ src/components/Hero/Hero.tsx
@@ -30,7 +30,10 @@ export const heroSecondaryTypes = [
   "secondaryWhatsOn",
 ];
 export interface HeroImageProps
-  extends Pick<ComponentImageProps, "alt" | "src"> {}
+  extends Pick<
+    ComponentImageProps,
+    "alt" | "fallbackSrc" | "src" | "onError"
+  > {}
 export interface HeroProps {
   /**
    * Optional background color for the backdrop only in the `campaign` variant.
@@ -55,13 +58,16 @@ export interface HeroProps {
   /** Used to control how the `Hero` component will be rendered. */
   heroType?: HeroTypes;
   /** Object used to create and render the `Image` component. Note that only
-   * `src` and `alt` are the available attributes to pass. If `imageProps.alt`
-   * is left blank, a warning will be logged to the console and will cause
-   * accessibility issues. For `imageProps.src`, it will only work for the
-   * "secondary", "fiftyFifty" and "campaign" `Hero` types; Note: `imageProps.src`
+   * `alt`, `component`, and `src` are the available attributes to pass. If
+   * `imageProps.alt` is left blank, a warning will be logged to the console and
+   * will cause accessibility issues. For `imageProps.src`, it will only work for
+   * the "secondary", "fiftyFifty" and "campaign" `Hero` types; Note: `imageProps.src`
    * can only be used in conjunction with `backgroundImageSrc` for the "campaign"
    * `Hero` type. Note: not all `Hero` variations utilize this prop. */
   imageProps?: HeroImageProps;
+  /** Optional boolean used to toggle the default text color from light to dark.
+   * Set isDarkText to `true` if the backgroundColor is set to a light color. */
+  isDarkText?: boolean;
   /** Optional boolean used to toggle the treatment of the background image in
    * the "campaign" variant. If true, the background image will be converted to
    * black & white and darkened to 60% black. */
@@ -93,11 +99,16 @@ export const Hero: ChakraComponent<
           alt: "",
           src: "",
         },
+        isDarkText,
         isDarkBackgroundImage = false,
         locationDetails,
         subHeaderText,
       } = props;
-      const styles = useMultiStyleConfig("Hero", { variant: heroType });
+      const styles = useMultiStyleConfig("Hero", {
+        foregroundColor,
+        isDarkText,
+        variant: heroType,
+      });
       const headingStyles = styles.heading;
       // We want to add `Hero`-specific styling to the `Heading` component.
       const finalHeading =
@@ -107,7 +118,8 @@ export const Hero: ChakraComponent<
 
       if (imageProps.src && !imageProps.alt) {
         console.warn(
-          `NYPL Reservoir Hero: The "imageProps.src" prop was passed but the "imageProps.alt" props was not. This will make the rendered image inaccessible.`
+          `NYPL Reservoir Hero: The "imageProps.src" prop was passed but the "imageProps.alt"` +
+            ` props was not. This will make the rendered image inaccessible.`
         );
       }
 
@@ -176,10 +188,10 @@ export const Hero: ChakraComponent<
        * into the component file and the related styles for all variants, other
        * than the "secondary" variant, were removed from the theme file. */
       const allDefaultBackgroundColors = {
-        primary: useColorModeValue("ui.bg.default", "dark.ui.bg.default"),
+        primary: useColorModeValue("ui.black", "dark.ui.bg.default"),
         secondary: useColorModeValue("ui.bg.default", "dark.ui.bg.default"),
         tertiary: useColorModeValue("ui.gray.x-dark", "dark.ui.bg.default"),
-        campaign: useColorModeValue("dark.ui.bg.default", "dark.ui.bg.default"),
+        campaign: useColorModeValue("ui.black", "dark.ui.bg.default"),
         campaignBackdrop: useColorModeValue(
           "dark.ui.bg.active",
           "dark.ui.bg.active"
@@ -258,7 +270,7 @@ export const Hero: ChakraComponent<
           ...(foregroundColor && { color: foregroundColor }),
           ...(backgroundColor
             ? { backgroundColor }
-            : { defaultBackgroundColor }),
+            : { bgColor: defaultBackgroundColor }),
         };
       } else if (
         foregroundColor ||
@@ -271,6 +283,22 @@ export const Hero: ChakraComponent<
             "`'secondary'` `heroType` variant will not use them."
         );
       }
+      if (foregroundColor && isDarkText) {
+        console.warn(
+          "NYPL Reservoir Hero: The `foregroundColor` and `isDarkText` props " +
+            "have both been passed. Thse props can not be used at the same time, " +
+            "so the `foregroundColor` prop will override the `isDarkText` prop."
+        );
+      }
+
+      const imageToRender = (
+        <Image
+          alt={imageProps.alt}
+          fallbackSrc={imageProps.fallbackSrc}
+          onError={imageProps.onError}
+          src={imageProps.src}
+        />
+      );
 
       const childrenToRender =
         heroType === "campaign" ? (
@@ -283,7 +311,7 @@ export const Hero: ChakraComponent<
                 backgroundImage: `/**/url("${imageProps.src}")`,
               }}
             >
-              <Image alt={imageProps.alt} src={imageProps.src} />
+              {imageToRender}
             </Box>
             <Box __css={styles.interior}>
               {finalHeading}
@@ -304,7 +332,7 @@ export const Hero: ChakraComponent<
                       : undefined,
                 }}
               >
-                <Image alt={imageProps.alt} src={imageProps.src} />
+                {imageToRender}
               </Box>
             )}
             {finalHeading}